refactor: 拆分 useChart 为三个独立 composable
- useChartCore.ts — 图表创建/销毁、样式主题、精度自适应，公共基础
- useChart.ts — K 线专用：懒加载历史数据、MA 叠加、副图指标切换
- useMinuteChart.ts — 分时图专用：分钟数据映射、增量刷新
- MinuteChart.vue 改为引用 useMinuteChart

拆分后职责清晰：KLineChart 和 MinuteChart 不再共用混乱的 useChart
